Nearly total rewrite #1 :)
After having learned lessons about `attrs` and also having realized I
was confused about the descriptor protocol vs. the `__getitem__`
container interface, I decided to restructure hermes from the ground
up.

Currently, the BaseTimeAccount tries to define as much as possible using
hopefully an air-tight unopinionated interface, however in a few places
it refers directly to TimeAccount. I'll need to address this eventually,
and either undo those references, or else make sure that the references
can be easily overwritten to point at a different reference or
something. But that's a problem for another day.