Switch to ortools for schedule solving
This is a fairly large change, and marks a huge milestone for hermes!

First, I've switched from `python-constraint` to `ortools`'s Cp-SAT
solver, which should be far more expressive, powerful, and fast.
`ortools` is maintained by Google, so hopefully that's a good sign?
`ortools` seems to have a bit of a 'generational rewrite' problem where
it seems that every few years they implement a "newer, better" version
and distribute it alongside the old version in the same package and
namespace, which is ugly, but the newest Cp-SAT solver seems to be well
designed with good documentation.

Second, and far more importantly - `schedule.py` now LITERALLY ACTUALLY
DOES THE THING! If you run `schedule.py` it will generate a working
schedule and upload it to google calendar!! HOLY COW!!

Right now the schedule is pretty dumb, but hey, it's a starting point!